Fix PHPStan errors from mis-inferred parameter types

+++ app/bundles/CoreBundle/Helper/ColorHelper.php
@@ -79,7 +79,7 @@ final class ColorHelper
      *
      * @param float $alpha (0 - 1)
      */
-    public function toRgba(int $alpha = 1): string
+    public function toRgba(float $alpha = 1): string
     {
         return sprintf('rgba(%d,%d,%d,%g)', $this->red, $this->green, $this->blue, (float) $alpha);
     }
